fpod refactor
